requirements.txt


cloudinary==1.29.0
Django==4.0.3
django-cloudinary-storage==0.3.0
django-ckeditor==6.4.2
django-environ==0.8.1
django-extensions==3.1.5
django-admin-interface==0.20.0
django-social-share==2.3.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
django-hitcount==1.3.5
django-taggit==3.0.0
django-crispy-forms==1.14.0
django-cors-headers
dj-database-url==0.5.0
dj-static==0.0.6
gunicorn==20.1.0
python-dotenv==0.20.0
static3==0.7.0
whitenoise==6.0.0
Pillow
django-widget-tweaks==1.4.12
